Refactor phone numbers validations to concern
